+++ tests/testthat/test-fact.R
@@ -246,6 +246,33 @@ test_that("drop_levels() works", {
 })
 
 
+# fact_reverse() ----------------------------------------------------------
+
+test_that("fact_reverse() works", {
+  res <- fact_reverse(1:4)
+  exp <- new_fact(4:1, 4:1)
+  expect_identical(res, exp)
+
+  res <- fact_reverse(as_ordered(1:4))
+  exp <- new_fact(4:1, 4:1, ordered = TRUE)
+  expect_identical(res, exp)
+
+  res <- fact_reverse(c(1:3, NA))
+  exp <- new_fact(c(3:1, 4L), c(3:1, NA))
+  expect_identical(res, exp)
+
+  res <- fact_reverse(as_ordered(c(1:3, NA)))
+  exp <- struct(
+    c(3:1, NA),
+    class = c("fact", "ordered", "factor"),
+    levels = c("3", "2", "1"),
+    uniques = 3:1,
+    na = 0L
+  )
+
+  expect_identical(res, exp)
+})
+
 
 # other methods -----------------------------------------------------------